requirements.txt
aiogram==3.20.0.post0
mysql-connector-python==9.3.0
python-dotenv==1.1.1